Move shared HTTP/retry logic to adapters/utils.py
Move ChatCompletionError, send_chat_request, error detection helpers,
and backoff logic from gateway_adapter.py to the shared utils module.
This positions the retry infrastructure for reuse when litellm adapter
is eventually consolidated.
