+++ mlflow/genai/judges/adapters/gateway_adapter.py
@@ -9,12 +9,10 @@ from __future__ import annotations
 
 import json
 import logging
-import time
 from dataclasses import dataclass
 from typing import TYPE_CHECKING, Any
 
 import pydantic
-import requests
 
 if TYPE_CHECKING:
     from mlflow.entities.trace import Trace
@@ -26,15 +24,16 @@ from mlflow.entities.assessment_source import AssessmentSource, AssessmentSource
 from mlflow.environment_variables import MLFLOW_JUDGE_MAX_ITERATIONS
 from mlflow.exceptions import MlflowException
 from mlflow.gateway.config import EndpointType
-from mlflow.gateway.constants import (
-    MLFLOW_GATEWAY_CLIENT_QUERY_RETRY_CODES,
-    MLFLOW_GATEWAY_ROUTE_TIMEOUT_SECONDS,
-)
 from mlflow.genai.discovery.utils import _pydantic_to_response_format
 from mlflow.genai.judges.adapters.base_adapter import (
     AdapterInvocationInput,
     AdapterInvocationOutput,
     BaseJudgeAdapter,
+)
+from mlflow.genai.judges.adapters.utils import (
+    ChatCompletionError,
+    is_response_format_error,
+    send_chat_request,
 )
 from mlflow.genai.judges.utils.parsing_utils import (
     _sanitize_justification,
@@ -65,14 +64,6 @@ class InvokeOutput:
     request_id: str | None
     num_prompt_tokens: int | None
     num_completion_tokens: int | None
-
-
-class ChatCompletionError(Exception):
-    def __init__(self, status_code: int, message: str, is_context_window_error: bool = False):
-        self.status_code = status_code
-        self.message = message
-        self.is_context_window_error = is_context_window_error
-        super().__init__(message)
 
 
 # ---------------------------------------------------------------------------
@@ -202,105 +193,6 @@ def _build_request(
         payload.update(inference_params)
 
     return payload
-
-
-def _send_chat_request(
-    endpoint: str,
-    headers: dict[str, str],
-    payload: dict[str, Any],
-    num_retries: int,
-) -> dict[str, Any]:
-    """Send a chat completions request with retry logic."""
-    last_exception = None
-    for attempt in range(1 + num_retries):
-        try:
-            resp = requests.post(
-                url=endpoint,
-                headers={"Content-Type": "application/json", **headers},
-                json=payload,
-                timeout=MLFLOW_GATEWAY_ROUTE_TIMEOUT_SECONDS,
-            )
-
-            if resp.status_code == 400:
-                # Don't retry 400s — raise immediately for caller to handle
-                error_body = _safe_parse_error(resp)
-                raise ChatCompletionError(
-                    status_code=400,
-                    message=error_body,
-                    is_context_window_error=_is_context_window_error(error_body),
-                )
-
-            if resp.status_code in MLFLOW_GATEWAY_CLIENT_QUERY_RETRY_CODES:
-                last_exception = ChatCompletionError(
-                    status_code=resp.status_code,
-                    message=_safe_parse_error(resp),
-                )
-                if attempt < num_retries:
-                    _sleep_with_backoff(attempt)
-                    continue
-                raise last_exception
-
-            if resp.status_code >= 400:
-                raise ChatCompletionError(
-                    status_code=resp.status_code,
-                    message=_safe_parse_error(resp),
-                )
-
-            return resp.json()
-
-        except requests.exceptions.Timeout as e:
-            last_exception = e
-            if attempt < num_retries:
-                _sleep_with_backoff(attempt)
-                continue
-            raise MlflowException(
-                f"Request to {endpoint} timed out after {MLFLOW_GATEWAY_ROUTE_TIMEOUT_SECONDS}s",
-                error_code=INTERNAL_ERROR,
-            ) from e
-        except requests.exceptions.ConnectionError as e:
-            last_exception = e
-            if attempt < num_retries:
-                _sleep_with_backoff(attempt)
-                continue
-            raise MlflowException(
-                f"Failed to connect to {endpoint}: {e}",
-                error_code=INTERNAL_ERROR,
-            ) from e
-
-    raise last_exception
-
-
-def _safe_parse_error(resp: requests.Response) -> str:
-    try:
-        body = resp.json()
-        if "error" in body:
-            error = body["error"]
-            if isinstance(error, dict):
-                return error.get("message", resp.text)
-            return str(error)
-    except Exception:
-        pass
-    return resp.text
-
-
-def _is_response_format_error(error_message: str) -> bool:
-    lower = error_message.lower()
-    return (
-        "response_format" in lower
-        or "json_schema" in lower
-        or "structured output" in lower
-        or "output_config" in lower
-    )
-
-
-def _is_context_window_error(error_message: str) -> bool:
-    lower = error_message.lower()
-    return "context length" in lower or "too many tokens" in lower or "maximum context" in lower
-
-
-def _sleep_with_backoff(attempt: int) -> None:
-    delay = min(2**attempt, 30)
-    time.sleep(delay)
 
 
 # ---------------------------------------------------------------------------
@@ -674,7 +566,7 @@ class GatewayAdapter(BaseJudgeAdapter):
                 )
 
                 try:
-                    response_data = _send_chat_request(
+                    response_data = send_chat_request(
                         endpoint=endpoint,
                         headers=headers,
                         payload=chat_payload,
@@ -695,7 +587,7 @@ class GatewayAdapter(BaseJudgeAdapter):
                     if (
                         e.status_code == 400
                         and include_response_format
-                        and _is_response_format_error(e.message)
+                        and is_response_format_error(e.message)
                     ):
                         _logger.debug(
                             f"Model {provider}/{model_name} may not support structured "
